refactor: global enum 클래스 파일 위치 변경
