+++ src/main/java/com/umc/networkingService/domain/inviteCode/entity/InviteCode.java
@@ -2,7 +2,7 @@ package com.umc.networkingService.domain.inviteCode.entity;
 
 import com.umc.networkingService.domain.member.entity.Member;
 import com.umc.networkingService.global.common.base.BaseEntity;
-import com.umc.networkingService.global.common.Role;
+import com.umc.networkingService.global.common.enums.Role;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
