+++ src/main/java/com/umc/networkingService/domain/member/entity/Member.java
@@ -3,12 +3,10 @@ package com.umc.networkingService.domain.member.entity;
 import com.umc.networkingService.domain.branch.entity.Branch;
 import com.umc.networkingService.domain.university.entity.University;
 import com.umc.networkingService.global.common.base.BaseEntity;
-import com.umc.networkingService.global.common.Part;
-import com.umc.networkingService.global.common.Role;
-import com.umc.networkingService.global.common.Semester;
+import com.umc.networkingService.global.common.enums.Part;
+import com.umc.networkingService.global.common.enums.Role;
+import com.umc.networkingService.global.common.enums.Semester;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
