+++ src/main/java/com/umc/networkingService/domain/project/entity/ProjectMember.java
@@ -2,7 +2,7 @@ package com.umc.networkingService.domain.project.entity;
 
 import com.umc.networkingService.domain.member.entity.Member;
 import com.umc.networkingService.global.common.base.BaseEntity;
-import com.umc.networkingService.global.common.Part;
+import com.umc.networkingService.global.common.enums.Part;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
