+++ src/main/java/com/umc/networkingService/global/common/enums/Part.java
@@ -1,4 +1,4 @@
-package com.umc.networkingService.global.common;
+package com.umc.networkingService.global.common.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
